x/codenet/types/message_encode_data_test.go
package types

import (
	"testing"

	"codenet/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgEncodeData_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgEncodeData
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgEncodeData{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgEncodeData{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
